requirements.txt
Django==2.0.7
djangorestframework==3.8.2
Pillow==5.2.0
pkg-resources==0.0.0
psycopg2==2.7.5
pytz==2018.5
